feat: Доделал 2 задание. Работает фильтрация, переделано на FSD, данные сохраняются через Zustand в local storage
