+++ src/app/App.tsx
@@ -1,6 +1,10 @@
 import { RouterProvider } from 'react-router-dom'
 import { router } from './router'
 
+/**
+ * Основной компонент приложения
+ * @returns 
+ */
 function App() {
   return (
     <RouterProvider router={router} />
