+++ src/entities/task/ui/TaskItem.tsx
@@ -5,7 +5,8 @@ import TaskStatus from './TaskStatus'
 import TaskPriority from './TaskPriority'
 import type { TContext } from '@/shared/types/types'
 import { Link } from 'react-router-dom'
-import { getCreatedTime } from '@/shared/lib/getCreatedTime'
+import { setCreatedTime } from '@/shared/lib/setCreatedTime'
+
 
 /**
   Карточка задачи, содержащая:
@@ -17,32 +18,35 @@ import { getCreatedTime } from '@/shared/lib/getCreatedTime'
 
 function TaskItem({ task }: { task: TContext }) {
     return (
-        <Card className="h-fit w-full max-w-xs mb-4 overflow-x-auto relative">
-            <CardHeader>
-                <CardTitle className="flex justify-between">
-                    <CardTitle>{task.header}</CardTitle>
-                    <CardTitle className="text-neutral-400">{getCreatedTime()}</CardTitle>
-                </CardTitle>
+        <Card className="h-fit w-full max-w-xs mb-4 overflow-x-auto relative 
+        hover:shadow-accent hover:shadow-md">
+            <Link to={`/task/${task.id}/edit`}>
+                <CardHeader>
+                    <CardTitle className="flex justify-between">
+                        <CardTitle>{task.header}</CardTitle>
+                        <CardTitle className="text-neutral-400">{setCreatedTime(task)}</CardTitle>
+                    </CardTitle>
 
-                <CardDescription className="wrap-break-word overflow-auto ">
-                    {task.description}
-                </CardDescription>
+                    <CardDescription className="wrap-break-word overflow-auto ">
+                        {task.description}
+                    </CardDescription>
 
-                <CardDescription>
-                    Приоритет:
-                    <TaskPriority priority={task.priority} />
-                </CardDescription>
+                    <CardDescription>
+                        Приоритет:
+                        <TaskPriority priority={task.priority} />
+                    </CardDescription>
 
-                <CardDescription>
-                    Категория:
-                    <TaskCategory category={task.category} />
-                </CardDescription>
+                    <CardDescription>
+                        Категория:
+                        <TaskCategory category={task.category} />
+                    </CardDescription>
 
-                <CardDescription>
-                    Статус:
-                    <TaskStatus status={task.status} />
-                </CardDescription>
-            </CardHeader>
+                    <CardDescription>
+                        Статус:
+                        <TaskStatus status={task.status} />
+                    </CardDescription>
+                </CardHeader>
+            </Link>
 
             <CardFooter className="flex flex-col mt-auto justify-center">
                 <Button variant="outline" className="hover:cursor-pointer w-full sm:w-auto">
