+++ src/entities/task/ui/TaskList.tsx
@@ -1,124 +1,45 @@
 import TaskItem from '../../../entities/task/ui/TaskItem'
 import { Button } from '@/shared/ui/button'
 import { Link } from 'react-router-dom'
-import type { TCategory, TPriority, TStatus } from '@/shared/types/types'
 import generateTaskName from '@/shared/lib/generateTaskName'
-import { useState } from 'react'
-import {
-    Select,
-    SelectContent,
-    SelectGroup,
-    SelectItem,
-    SelectLabel,
-    SelectTrigger,
-    SelectValue,
-} from '@/shared/ui/select'
-import { useStore } from '@/app/store'
+import { useStore } from '@/app/stores/store'
+import FilterTask from '@/features/task/filter-task'
+import { useFilterStore } from '@/app/stores/useFilterStore'
 
 /** 
 Отображает список задач в виде адаптивной сетки карточек (Grid или Flexbox).
-
 Каждая задача отображается через компонент TaskItem
 */
 
 function TaskList() {
-    // const { state } = useContext<TContextType>(Context)
     const tasks = useStore((state) => state.tasks)
-
-    const [statusFilter, setStatusFilter] = useState<string>('')
-    const [categoryFilter, setCategoryFilter] = useState<string>('')
-    const [priorityFilter, setPriorityFilter] = useState<string>('')
-
-    const statusList: Array<TStatus> = ['To Do', 'In Progress', 'Done']
-    const categoryList: Array<TCategory> = ['Bug', 'Feature', 'Documentation', 'Refactor', 'Test']
-    const priorityList: Array<TPriority> = ['Low', 'Medium', 'High']
+    const { status, category, priority } = useFilterStore();
 
     // Фильтрация задач
     const filteredTasks = tasks.filter(
         (tasks) =>
-            (!statusFilter || tasks.status === statusFilter) &&
-            (!categoryFilter || tasks.category === categoryFilter) &&
-            (!priorityFilter || tasks.priority === priorityFilter)
+            (!status || tasks.status === status) &&
+            (!category || tasks.category === category) &&
+            (!priority || tasks.priority === priority)
     )
-
-    // Сброс фильтров
-    const resetFilters = () => {
-        setStatusFilter('')
-        setCategoryFilter('')
-        setPriorityFilter('')
-    }
-
     return (
         <section className="min-h-screen w-full flex flex-col box-border items-center justify-end bg-neutral-200 overflow-auto p-2 sm:p-4">
             <div className="flex flex-col sm:flex-row w-full mb-2 gap-2 bg-neutral-200 items-center justify-between p-2 sm:p-4">
                 <h1 className="text-lg sm:text-2xl font-bold">Task Bar</h1>
-                <div className="flex flex-col sm:flex-row gap-2 items-center">
-                    {/* Фильтр по приоритету */}
-                    <Select onValueChange={(e) => setPriorityFilter(e)} value={priorityFilter}>
-                        <SelectTrigger className="w-[200px] bg-white">
-                            <SelectValue placeholder="Установите приоритет" />
-                        </SelectTrigger>
-                        <SelectContent>
-                            <SelectGroup>
-                                <SelectLabel>Приоритет</SelectLabel>
-                                {priorityList.map((priority) => (
-                                    <SelectItem key={priority} value={priority}>
-                                        {priority}
-                                    </SelectItem>
-                                ))}
-                            </SelectGroup>
-                        </SelectContent>
-                    </Select>
-
-                    {/* Фильтр по категории */}
-                    <Select onValueChange={(e) => setCategoryFilter(e)} value={categoryFilter}>
-                        <SelectTrigger className="w-[200px] bg-white">
-                            <SelectValue placeholder="Установите категорию" />
-                        </SelectTrigger>
-                        <SelectContent>
-                            <SelectGroup>
-                                <SelectLabel>Категория</SelectLabel>
-                                {categoryList.map((category) => (
-                                    <SelectItem key={category} value={category}>
-                                        {category}
-                                    </SelectItem>
-                                ))}
-                            </SelectGroup>
-                        </SelectContent>
-                    </Select>
-
-                    {/* Фильтр по статусу */}
-                    <Select onValueChange={(e) => setStatusFilter(e)} value={statusFilter}>
-                        <SelectTrigger className="w-[200px] bg-white">
-                            <SelectValue placeholder="Установите статус" />
-                        </SelectTrigger>
-                        <SelectContent>
-                            <SelectGroup>
-                                <SelectLabel>Статус</SelectLabel>
-                                {statusList.map((status) => (
-                                    <SelectItem key={status} value={status}>
-                                        {status}
-                                    </SelectItem>
-                                ))}
-                            </SelectGroup>
-                        </SelectContent>
-                    </Select>
-
-                    <Button
-                        onClick={resetFilters}
-                        className="hover:cursor-pointer box-border border-2 border-chart-2"
-                        variant={'outline'}
-                    >
-                        Сбросить фильтры
-                    </Button>
-
+                <div className="flex flex-col gap-2 bg-neutral-200
+                lg:flex-row
+                lg:justify-center
+                sm:gap-4">
+                    {/* Компонент с фильтрами */}
+                    <FilterTask />
+                    {/* Компонент с фильтрами */}
                     <Button className="hover:cursor-pointer" variant={'default'}>
                         <Link to={`task/${generateTaskName().number}`}>Создать задачу</Link>
                     </Button>
                 </div>
             </div>
             <div className="flex flex-col w-full flex-1 bg-neutral-200 border-2 border-neutral-500 rounded-2xl p-2 sm:p-4">
-                <h1 className="text-base sm:text-xl font-semibold mb-2">Task List</h1>
+                {/* <h1 className="text-base sm:text-xl font-semibold mb-2">Task List</h1> */}
                 <div
                     className="grid grid-cols-1 gap-4
                 sm:grid-cols-2 max-sm:place-items-center max-sm:gap-2
